requirements.txt
httpx>=0.27.0
python-telegram-bot>=21.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
